+++ app/components/graph/chart-container-service.js
@@ -31,6 +31,7 @@ angular.module('lizard-nxt')
     this.keys = chartContent.keys || defaultKeys;
     this.color = chartContent.color || DEFAULT_GREEN;
     this.unit = chartContent.unit;
+    this.thresholds = chartContent.thresholds;
     this.location = chartContent.location;
     this.labels = chartContent.labels;
     this.setContentUpdateY(chartContent);